requirements.txt
flask
pathway
feedparser
deepface
opencv-python
numpy
Pillow
requests
openai
rapidfuzz
python-dotenv
Werkzeug
